Add inlineSearch feature
If inlineSearch is enabled, search bar will be added to top of PodcastList of DiscoveryResults. It will be disabled by default.
SearchBar will behave similarly to Search icon on app bar except the search term can now be passed from the to Search view.
Set autoFocus of search TextField in Search view to false when the search term is passed.

Use Icon color from theme